Make render_room_row_minimap accept Vec2f instead of Vec2i

+++ src/something_main.cpp
@@ -116,7 +116,7 @@ void render_debug_overlay(Game_State game_state, SDL_Renderer *renderer)
              entities[PLAYER_ENTITY_INDEX].vel.x,
              entities[PLAYER_ENTITY_INDEX].vel.y);
 
-    const auto minimap_position = vec2((int)PADDING, (int)(6 * 50 + PADDING));
+    const auto minimap_position = vec2(PADDING, 6 * 50 + PADDING);
     render_room_row_minimap(renderer, minimap_position);
     render_entity_on_minimap(
         renderer,
